requirements.txt
appdirs==1.4.2
numpy==1.12.0
olefile==0.44
opencv-python==3.2.0.6
packaging==16.8
pandas==0.19.2
Pillow==4.0.0
protobuf==3.2.0
pydicom==0.9.9
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
six==1.10.0
tensorflow==1.0.0
